Enable PhantomJS disk cache
Unfortunately there is a bug in PhantomJS where redirects and caching down play
nice together and result in the Poltergeist/PhantomJS forgetting all about the
page it was supposed to have just loaded and reporting a blank document.

Because this bug exists only in the in-memory cache, we can just enable the disk
cache to get around it. Eventually when people's OS package managers begin
shipping versions of PhantomJS where this issue is fixed. For now, Ubuntu is too
far behind, so we might as well leave this on this setting which should work for
all users.

Bug: https://github.com/teampoltergeist/poltergeist/issues/754